User input filtering

+++ lib/models/lhgallery/erlhcoreclassmodellastsearch.php
@@ -21,33 +21,34 @@ class erLhcoreClassModelGalleryLastSearch {
    
    public static function addSearch($keyword,$search_count)
    {
-       $db = ezcDbInstance::get('slave');
-       $stmt = $db->prepare('SELECT count(id) FROM `lh_gallery_lastsearch` WHERE keyword = :keyword');  
-       $stmt->bindValue( ':keyword',$keyword);
-       $stmt->execute(); 
-       $count = $stmt->fetchColumn(); 
+       
+       if (strpos($keyword,'%25') === false){
+           $db = ezcDbInstance::get('slave');
+           $stmt = $db->prepare('SELECT count(id) FROM `lh_gallery_lastsearch` WHERE keyword = :keyword');  
+           $stmt->bindValue( ':keyword',$keyword);
+           $stmt->execute(); 
+           $count = $stmt->fetchColumn(); 
     
-       erLhcoreClassModelGallerySearchHistory::addSearch($keyword,$search_count);
-    
-       if  ($count == 0     )
-       {       
-           try {
-               $search = new erLhcoreClassModelGalleryLastSearch();
-               $search->keyword = $keyword;
-               $search->countresult = $search_count;       
-               erLhcoreClassGallery::getSession()->save($search);
-           } catch (Exception $e) { // Sometimes table gets crushed if a lot of searches is done
+           if  ($count == 0     )
+           {       
+               try {
+                   $search = new erLhcoreClassModelGalleryLastSearch();
+                   $search->keyword = $keyword;
+                   $search->countresult = $search_count;       
+                   erLhcoreClassGallery::getSession()->save($search);
+               } catch (Exception $e) { // Sometimes table gets crushed if a lot of searches is done
+                   
+               }
                
+               $db = ezcDbInstance::get();
+               $stmt = $db->prepare('SELECT id FROM `lh_gallery_lastsearch` order by id desc limit 9,1');    
+               $stmt->execute();
+               $idlast = $stmt->fetchColumn();        
+                        
+               $stmt = $db->prepare('DELETE FROM `lh_gallery_lastsearch` WHERE id < :id_last'); 
+               $stmt->bindValue( ':id_last',$idlast);  
+               $stmt->execute();    
            }
-           
-           $db = ezcDbInstance::get();
-           $stmt = $db->prepare('SELECT id FROM `lh_gallery_lastsearch` order by id desc limit 9,1');    
-           $stmt->execute();
-           $idlast = $stmt->fetchColumn();        
-                    
-           $stmt = $db->prepare('DELETE FROM `lh_gallery_lastsearch` WHERE id < :id_last'); 
-           $stmt->bindValue( ':id_last',$idlast);  
-           $stmt->execute();    
        }
    }
